Fix various warnings.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4353 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/huff.c
@@ -26,8 +26,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #ifdef HUFFNETWORK
 #define ID_INLINE
 
-#define VALUE(a)			(*(int  *)&(a))
-#define NODE(a)				((void*)(a))
+#define VALUE(a)			(*(size_t *)&(a))
+#define NODE(a)				((void*)((size_t)a))
 
 #define NODE_START			NODE(  1)
 #define NODE_NONE			NODE(256)
